# Conflicts:
#	examples/reshare/src/engine.rs

+++ examples/sync/src/net/mod.rs
@@ -3,7 +3,7 @@ use commonware_codec::{DecodeExt, Encode, EncodeSize, Error, Read, ReadExt, Read
 use std::mem::size_of;
 
 /// Maximum message size in bytes (10MB).
-pub const MAX_MESSAGE_SIZE: usize = 10 * 1024 * 1024;
+pub const MAX_MESSAGE_SIZE: u32 = 10 * 1024 * 1024;
 
 pub mod request_id;
 pub use request_id::RequestId;
@@ -100,7 +100,7 @@ impl Read for ErrorResponse {
     fn read_cfg(buf: &mut impl Buf, _: &()) -> Result<Self, Error> {
         let request_id = RequestId::read_cfg(buf, &())?;
         let error_code = ErrorCode::read(buf)?;
-        let message_bytes = Vec::<u8>::read_range(buf, 0..=MAX_MESSAGE_SIZE)?;
+        let message_bytes = Vec::<u8>::read_range(buf, 0..=MAX_MESSAGE_SIZE as usize)?;
         let message = String::from_utf8(message_bytes)
             .map_err(|_| Error::Invalid("ErrorResponse", "invalid UTF-8 in message"))?;
         Ok(Self {
